Remove check for 9600 baud.
It is true that 9600 baud is better (see "It's the Latency, Stupid"
by Stuart Cheshire) but refusing to even start a game if the wrong
connect string is detected is very inconvenient. The later versions
of sersetup ditched this check anyway, and DOSbox's soft modem
always connects at 57600 baud, making sersetup impossible to test in
DOSbox unless this check is removed.