metanet: Add initial node discovery code.
On startup we must discover all nodes in the network. To achieve this,
start from the initial set of immediate neighbor nodes that we are
aware of via underlying drivers. Once a second, send a discover
message to all nodes we are aware of that includes information about
immediate neighbors. When receiving a discover message we use this
information to increase the size of our set.