metanet: Combine identical packets into broadcasts.
Doom assumes any underlying network driver is using a direct peer-to-peer
link, and this means that network traffic can grow with O(N^2) of the
number of players in the game. With metanet this can swamp bottleneck
links like slow serial connections; with my tests using 8-player Hexen
bottlenecked on a simulated serial link, I was unable to get the game to
even start unless -dup 4 was specified.

This uses heuristics based on knowledge of the NetUpdate() function in the
Doom source to detect when the exact same packet is sent to every node in
sequence, and send such packets only once as a broadcast through the
network. There are some degenerate cases where it can fall back to sending
individual packets if retransmits are requested, but in practice it seems
to work well and the majority of transmits are upgraded to broadcasts.
In my test I can now start and run 8-player Hexen with -dup 1 without any
problems.