Add Analog Joystick driver.
analogjs allows more precise control when using a PC joystick, by using
a linear scaling of stick movement to player movement and turning. This is
nice when using a modern controller with analog sticks (and the game
running in DOSBox).

Assembler is from the Heretic sources and adapted to 16-bit.