Get rid of user_data from launch APIs.
This is good practice for most C APIs as it allows context to be kept
when multiple callbacks of the same type may be invoked. However, in
this case it's pointless since when launching the game only a single
callback can ever be invoked. It's also unused in any of the programs
and complicates the API unnecessarily.