metanet: Use APIPointerFlag() to detect -net.
metanet needs to read multiple -net arguments on the command line
corresponding to the underlying drivers it will use for communications.
So use the new API pointer support to do this instead of parsing
command line arguments explicitly. This allows the code to be simplified
but also means that metanet no longer needs to be given a dummy command
to execute in forwarder mode.