Various MSC-related fixes.
These are from a (likely futile) attempt to compile the codebase with
Microsoft C 5.1. Although I'm unlikely to continue this attempt, these
are at least some worthwhile fixes - for example, I previously fixed a
bug with Turbo C when taking the address of an array using &, and this
eliminates many such cases.