net: Add ROTTCOM adapter driver.
This allows Rise of the Triad to run on top of Doom network drivers.
There is a slight gotcha in the fact that the ROTTCOM tx/rx buffer is
much larger than the Doom one, so we have to do packet fragmentation
and reassembly. The current code has a hard-coded hack to always
transmit to node 1, which works for two-player games; I haven't figured
out yet why ROTT is always trying to send to node 0.