Make callbacks invoked in ISRs far pointers.
We cannot rely on invoking near function pointers inside ISRs since the
segment pointers may be completely different. This has become apparent
as ef7568a9627db62a broke the network drivers, causing any invocations
of the network ISR to just hang. So fix this but also fix the control
API as well while we're at it since it uses the same mechanism.