sersetup: Use newer negotiation protocol.
Versions of SERSETUP after Doom v1.2 switched to this different player
negotiation protocol where a random number is used from both sides to
choose which player is which. Switch to this by default while keeping
compatibility with the old protocol in case it's really necessary for
some purpose. This also allows implementation of the -player2 flag.