ipxsetup: Add support for xttl setup extensions.
xttl's version of ipxsetup extends the setup info to check ticdup
matches between nodes and allow the player number to be explicitly
set. So add support for these in a cross-compatible way, with a
couple of additional enhancements:

* -dup not matching between nodes is not considered an error; if all
  nodes are using the vanilla-utilities version, they will adjust to
  settle on the largest value used by any node.
* Not all players need to specify -player and nodes that do not
  specify the player number will be auto-assigned.

Thanks to Danfun64 for prompting me to make sure that I reach feature
parity with xttl's ipxsetup.