Fix stack overflow crash with stack switching.
OpenWatcom includes a call to __STK in function headers that checks there
is sufficient stack space remaining without blowing the stack, but it
checks against the _STACKLOW internal variable which is set by the
compiler. When switching stacks we need to also update this variable,
otherwise we can erroneously trigger the stack overflow detection.