Add interrupt functions to save/restore PSP.
Certain DOS APIs, such as those to read from files, rely on data read
from the Program Segment Prefix (PSP). Inside an ISR, the PSP will be
set to that of the interrupted process, not our program. So add some
functions to allow us to temporarily switch back the PSP during our
ISRs, like we do with the stack.