Add SquashToResponseFile() to handle long command lines.
DOS has an excessively restrictive limit on the length of command
lines, and response files are the solution that Doom supports to work
around the problem. Detect command lines that exceed the limit and
automatically move out the minimum set of command line arguments to a
temporary response file. We delete the file on exit to clean up.

Since we also transparently read and expand response file arguments
(commit 06037a019c161391e), this means that the different tools should
just work together as intended.

Fixes #8.