ipxsetup: Add support for redirector-basd API.
There are two ways of invoking the Novell IPX API: the older API is
interrupt-based, while the newer API uses the DOS redirector API to
get a pointer to an entrypoint function to call. The original version
of ipxsetup (from which our version is derived) used the older API but
this was switched to the newer API subsequently. It's not clear why
(better performance? or to mitigate problems on some installations) but
it does suggest that we should probably also switch to it as well.

Later versions of the Novell suite apparently ditched the old API
entirely so we cannot rely on either API being present on all versions.
Therefore we try to detect both and set up a function pointer that
either calls OldIPXCall or NewIPXCall depending on what works.