Add support for "baked in" command line arguments.
The way that our versions of ipxsetup, sersetup, etc. work is more
flexible since the binary to run is not hard-coded like it is in the
originals. However, there are still times when it is convenient to
sometimes have versions that act like the originals. An example is if
we're using an old tool like DeathManager that expects the original
setup .exes. This "baked in arguments" feature is intended to allow our
binaries to be modified after build time to set arguments that are
always implicitly included when the program is run.