Delete code for old IPX API calls.
The approach of trying to call the GetLocalAddress API doesn't work
and in DOSbox can cause the emulator to just hang. I can't find any
example of how to detect IPX without using the redirector API (checked:
Quake, Descent, even the older version of the IPXSETUP code). So let's
just delete the code for the old calls and always use the new API,
since it's faster anyway.