parsetup: Add multiple receive buffers.
Previously we just had a single buffer for receiving packets, but it is
clear that the calls into the driver from doom.exe are not regular
enough to always receive packets before they got overwritten by the next
one. Instead we add a ring of receive buffers so that there is enough
breathing room before we start dropping packets. This appears to
eliminate most dropped packets.