metanet: Cease forwarding once all players quit.
When in packet forwarding mode we just forward packets without running
any other program. Previously we just did this in an infinite loop;
however, by sending a "quit" packet when nodes exit, we can detect when
there are no more nodes to be forwarded to and exit the loop
automatically.