Add new library functions for IRQ handlers.
The SERSETUP and PARSETUP programs both set up interrupt handlers to
send and receive data over ports, but it is done in a "raw" way rather
than through any kind of well-considered interface.

This is part of #12, though more work is needed for a full fix for that
bug so that different instances can share IRQs. The first code for ISR
chaining is in place here but is not yet invoked.