Add chaining for IRQ handlers.
This fixes #12. For the IRQ handlers installed by SERSETUP and PARSETUP,
we want to be able to share the IRQ when another instance of the same
binary is using the same IRQ. An example is that COM1 and COM3 usually
use the same IRQ. We can share the IRQ by chaining to the other ISR, but
we only want to do this when it's another vanilla-utilities binary -
usually we want the hardware all to ourselves.

We can identify this specific corner case by setting an environment
variable the first time we hook an IRQ. If another SERSETUP/PARSETUP
wants to use the same IRQ, it will notice the environment varialbe and
know that it has to chain.