parsetup: Disable interrupts on shutdown.
On startup we were enabling interrupts via the parallel port ACK line,
but we weren't undoing this and turning it off again on shutdown. This
seems like a sensible thing to do, since I'm not sure what even happens
if interrupts are allowed to continue after the ISR has been unloaded.