metanet: Increase station ID to 32-bits.
I occasionally see test failures when running the new test script
because of duplicate station IDs being detected. This is presumably an
instance of the birthday paradox, so let's bump up the station ID size
to 32-bits to mitigate the problem. Even if this isn't the cause,
using a larger station ID as it will reduce the probability of a clash
in real usage and there's no real reason not to do it.