sersetup: Refactor hardware code interface.
This is a first step to adding support for serial infrared mode (#18).
The previous version of the code was just "passive" - putting data into
and out of buffers and letting the code in serport.c take care of
transmit and receive. For infrared we want something a bit more direct
in terms of callbacks that can make decisions within the serial ISR. So
instead, add callbacks that are invoked to deliver bytes and get new
bytes to put into the UART transmit buffer.

We keep the ring buffer code that was previously part of serport.c, but
move it into sersetup.c, which in all honesty seems like a cleaner
separation anyway.