sersetup: Pause receive ISR calls on overload.
If we run out of receive buffer space, there's no sense in getting any more
receive notifications from the UART until the situation is remedied. This is
useful in the case of CPU overload where we might not actually be able to
receive data fast enough from the UART to keep up.

This might sound like a strangely hypothetical problem to address but is
actually a legitimate problem. When Doom starts up, player 1 does the "Sending
network start info..." step, transmitting the game settings to all other
players, but a bug in the code means that it actually *floods* all other
players by repeatedly sending the start packet over and over. During testing
on a slightly underpowered machine that struggled to run two dosbox instances
at once, player 2 would end up in a "wedged" state where the ISR did not
appear to be invoked any more. I was not able to conclusively determine why,
but turning off receive in the overload scenario fixes the problem.

Because of the same bug, this should speed up startup for player 2 on slow
machines, since the ISR will eventually stop being called, allowing the CPU to
focus on finishing R_Init().

Another step towards #18.