Call DOS idle functions in CheckAbort().
If we're calling CheckAbort() it means we're in a loop waiting -- in all
current cases for network input, though we're also listening for ESC being
pressed so that we can abort. Since we don't have any actual CPU work to do,
it's appropriate to give it up to other tasks on the system rather than just
spinning the CPU and doing nothing productive.