w9xsetup: Add Winsock2 VxD patcher program.
This applies a hotpatch to the running kernel that fixes an internal
table inside WSOCK2.VXD that contains the wrong values. Without this
patch applied it's not possible to use the Win9x DOS networking
backdoor when Winsock2 is installed.

There are some details about this bug at the top of the new file, but
some context can be found in Berczi Gabor's ws2dos.txt which gives
some details about the bug.