winsock: Fix for Winsock2 ioctlsocket() calls.
The Winsock2 ioctl function looks more like WSPIoctl() than the original
ioctlsocket() function. This took a bit of reverse engineering of the
msafd.dll library to figure out how it calls into wsock2.vxd; it looks
like the main thing that's missing was that the buffer size field was
not being initialized.

I'm still not entirely sure it's working as intended, but calls to
WS_ioctlsocket() to set a socket as nonblocking now no longer fail with
an error.