winsock: Perform API check for Winsock2 bug.
If we're using Winsock2 we need to check the bugfix has been applied.
So check that it has by creating a socket and trying to send a packet.
If the send fails with WSAENOTSOCK, we know that the patch needs to be
applied by running WS2PATCH. It's better to provide a useful error
message rather than a mysterious error code.