ipxsetup: Fix -port flag to take unsigned value.
This is an old bug that dates all the way back to the original IPXSETUP.
IPX port values are unsigned, but were being treated as signed when the
parameter to -port was parsed. This meant that large port numbers (>=
32768) would silently overflow.

Because the `port` variable was signed, there was an additional bug that
the default was initialized to the wrong value because it overflowed at
compile time. As a result, IPXSETUP would use a default value (0x869d)
that didn't actually match the "official" port number that was assigned
by Novell.

In fixing this I'm maintaining the "wrong" default port number because
compatibility is more useful than correctness and nobody really cares
about IPX port numbers not conflicting.