doomnet: Don't modify `args` argument.
This is not guaranteed to point to the start of something allocated
with `malloc` and in the case of udpsetup was not, causing some memory
corruption. Rather, we can use the already existing flag functions to
duplicate the list within `NetLaunchDoom()` so that it can mutate the
contents of the list however it wants.