winsock: Gate init on VxdGetEntryPoint, not VXDLDR.
Technically, even if we fail to use VXDLDR to load one of the WSOCK
VxDs, we might still succeed when we go to look up the VxD entrypoint
by its ID. An example is on Windows 3.x where the driver is named
WSOCK.386. So even if the VXDLDR part of initialization fails, carry
on anyway and try.