dossock: Use the correct signature for SOCKETS.EXE.
I was confused by the fact that Ralf Brown's interrupts doc specified
"TC" as the signature while MSSOCKS.PAS was using "SO". Disassembly of
the SDK reveals that "SO" is correct. Actually, it seems like both are
in a sense correct and the ioctl acts as a kind of redirector to select
the particular driver module to use. We want "SO" because we want the
sockets interface.