dossock: Pass parameter block pointer correctly.
MSSOCKS.PAS does something with the es:bx registers but I think this
is maybe something to do with making the call from protected mode into
the real mode driver. In reality the entrypoint seems to be a normal
stdcall (far) function pointer that takes a pointer to the parameters
block.