Change EndOfIRQ() function into a macro.
_chain_intr() seems to be a finicky beast, or at least "clever" in how
it works; I think it only works if you call it directly from the
original interrupt function. As such we need to implement this as a
macro.

Fixes a bug with sersetup where the Turbo C version would lock up when
two sersetups were sharing an interrupt. Actually, I'm not sure how the
Watcom version worked until now either.