metanet: Change type to work around Turbo C quirk.
Turbo C doesn't like a far array type as a function parameter, or at
least treats it differently from a far pointer. It gives warnings at
compile time but what it's clearly not working properly either. So
switch to a far pointer type to be more explicit about what we're trying
to do.