sirsetup: Expand handoff to be per-player.
Previously we had a single handoff character that indicated the other
side should take over, but if we want to support more than two players
this isn't sufficient. Instead, extend it to a new scheme where the
handoff peer indicates which player it's handing off to. Then we add the
concept of a "handoff partner" which is the player that each player will
hand off to after transmitting its packets; currently this is just hard-
coded to be 1 - consoleplayer, which works for two players.