sirsetup: Add source, destination addresses to header.
These will be used as the addressing scheme once we support more than
two players. For now we're still hard-coded to only support two players,
although we're already doing the internal steps of setting destination
and source fields correctly and setting doomcom.remotenode on delivery
based on the src field.