sirsetup: Detect more than two players.
This arbitration code is the main part missing to support more than two
players, adapted from the IPXSETUP code. There are a number of changes
from the IPXSETUP code to simplify things and to adapt to the specifics
of the SIR case. The -node parameter is not yet enabled since a few more
changes are still necessary.