Save bp, di registers during IPX call.
This fixes #21, where ipxsetup was not working properly with the Novell
IPXODI driver. Previous code (eg. in v1.3.0) worked but it seems to
have been coincidental because the code did not use many local
variables. The refactor to add udpsetup led to new code that assumed
particular values in the bp and di registers were preserved and broke
when they were clobbered by the IPX API. The same bug was not found
with Windows's IPX driver or DOSbox, presumably because they preserve
these registers.

At the same time, we don't need to preserve cx or dx - the stdcall
convention already assumes that the function may clobber them.