sirsetup: Don't drop empty packets
Vanilla Doom has a bug where it will send a flood of zero-tic packets
when it gets stuck. Nonetheless, it *is* necessary to send periodic
packets even if there are no tics in them, because retransmit only gets
triggered by a packet being received. Otherwise a deadlock can occur
where both sides are waiting for tics and cannot make progress.